% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRef}
\alias{scRef}
\title{Single cell obj for deconvolution.
The files were integrated with Seurat using healthy BM data from Triana et al, 2022 and Van Galen et al. 2020.}
\format{
An object of class \code{ExpressionSet} with 439 rows and 69101 columns.
}
\usage{
data(scRef)
}
\description{
Single cell obj for deconvolution.
The files were integrated with Seurat using healthy BM data from Triana et al, 2022 and Van Galen et al. 2020.
}
\examples{
data(scRef)
}
\concept{datasets}
\keyword{datasets}
